+++ src/index.ts
@@ -4,5 +4,6 @@ export * from "./executor.js";
 export * from "./json-schema.js";
 export * from "./policy.js";
 export * from "./registry.js";
+export * from "./report.js";
 export * from "./sarif.js";
 export * from "./schema.js";